% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRR-package.R
\docType{package}
\name{FAIRR-package}
\alias{FAIRR}
\alias{FAIRR-package}
\title{FAIRR: Functional Approximation of Impulse Responses}
\description{
Implementation of the Functional Approximation of Impulse Responses (FAIR) methodology. This package allows for the estimation of symmetric and asymmetric impulse response functions by approximating them as a linear combination of Gaussian basis or exponential basis functions using Stan.

Functional Approximation of Impulse Responses on R.
}
\references{
\itemize{
\item Barnichon, R., & Matthes, C. (2018). Functional Approximation of Impulse Responses. Journal of Monetary Economics, 99, 41-55.
\item Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
}
\author{
\strong{Maintainer}: Gauthier Serres \email{gauthierserres@outlook.fr}

}
\keyword{internal}
